requirements.txt
imbalanced_learn==0.12.2
imblearn==0.0
matplotlib==3.8.4
numpy==2.0.1
pandas==2.2.2
pytorch_tabular==1.1.0
PyYAML==6.0.2
scikit_learn==1.4.2
scipy==1.14.0
seaborn==0.13.2
xgboost==2.1.1
